+++ essrpc_macros/src/lib.rs
@@ -6,16 +6,34 @@ extern crate proc_macro2;
 extern crate quote;
 extern crate syn;
 
-use proc_macro::TokenStream;
+use core::convert::AsRef;
+use proc_macro::{TokenStream};
 use proc_macro2::TokenStream as TokenStream2;
-use proc_macro2::{Ident, Span};
-use quote::quote;
+use proc_macro2::{Ident, Span, TokenTree};
+use quote::{quote, ToTokens};
+use std::ops::Deref;
+use syn::parse_quote;
 use syn::{FnArg, ItemTrait, LitStr, Pat, TraitItem, TraitItemMethod};
 
 #[proc_macro_attribute]
 pub fn essrpc(args: TokenStream, input: TokenStream) -> TokenStream {
     // We don't handle any arguments today, perhaps we will in the future.
-    let _args = args.to_string();
+    let args: TokenStream2 = args.into();
+    let mut sync_client = false;
+    let mut async_client = false;
+    for tok in args {
+        if let TokenTree::Ident(ident) = tok {
+            match ident.to_string().as_ref() {
+                "sync" => sync_client = true,
+                "async" => async_client = true,
+                _ => ()
+            }
+        }
+    }
+
+    if !sync_client && !async_client {
+        sync_client = true
+    }
 
     let mut result: TokenStream2 = input.clone().into();
 
@@ -33,7 +51,13 @@ pub fn essrpc(args: TokenStream, input: TokenStream) -> TokenStream {
         }
     }
 
-    result.extend(create_client(&trait_ident, &methods));
+    if async_client {
+        result.extend(create_async_client_trait(&trait_ident, &methods));
+        result.extend(create_client(&async_client_trait_ident(&trait_ident), &methods, true));
+    }
+    if sync_client {
+        result.extend(create_client(&trait_ident, &methods, false));
+    }
     result.extend(create_server(&trait_ident, &methods));
 
     result.into()
@@ -41,6 +65,20 @@ pub fn essrpc(args: TokenStream, input: TokenStream) -> TokenStream {
 
 fn client_ident(trait_ident: &Ident) -> Ident {
     Ident::new(&format!("{}RPCClient", trait_ident), Span::call_site())
+}
+
+fn client_transport_ident(async_client: bool) -> Ident {
+    Ident::new(if async_client {"AsyncClientTransport"} else {"ClientTransport"},
+               Span::call_site())
+}
+
+fn async_client_trait_ident(trait_ident: &Ident) -> Ident {
+    Ident::new(&format!("{}Async", trait_ident), Span::call_site())
+}
+
+fn rpcclient_ident(async_client: bool) -> Ident {
+    Ident::new(if async_client {"AsyncRPCClient"} else {"RPCClient"},
+               Span::call_site())
 }
 
 fn server_ident(trait_ident: &Ident) -> Ident {
@@ -58,23 +96,31 @@ fn make_ident_literal_str(ident: &Ident) -> LitStr {
     let as_str = format!("{}", ident);
     LitStr::new(&as_str, Span::call_site())
 }
-    
 
-fn impl_client_method(method: &TraitItemMethod, id: u32) -> TokenStream2 {
+// True if has self param, false if has default implementation. Panics
+// if no self and no default.
+fn verify_self_param_or_unneeded(method: &TraitItemMethod) -> bool {
+    let param_tokens = &method.sig.decl.inputs;
+    let first = param_tokens.first();
+    if !first.is_some()
+        || (match first.unwrap().value() {
+            FnArg::SelfRef(_) => false,
+            _ => true
+        }) {
+            if method.default.is_some() {
+                // this method is not needed for the RPC client
+                return false
+            }
+            
+        }
+    return true;
+}
+
+// Client method implementation for the call to tx_begin_call through
+// tx_finalize. This portion is shared between sync and async.
+fn client_method_tx_send(method: &TraitItemMethod, id: u32) -> TokenStream2 {
     let ident = &method.sig.ident;
     let param_tokens = &method.sig.decl.inputs;
-
-    let first = param_tokens.first();
-    if !first.is_some() || (match first.unwrap().value() {
-        FnArg::SelfRef(_) => false,
-        _ => true
-    }) {
-        if method.default.is_some() {
-            // this method is not needed for the RPC client
-            return TokenStream2::new();
-        }
-        panic!("RPC methods must take '&self' as the first parameter, {} does not", ident);
-    }
 
     let mut add_param_tokens = TokenStream2::new();
 
@@ -86,20 +132,38 @@ fn impl_client_method(method: &TraitItemMethod, id: u32) -> TokenStream2 {
                 quote!(tr.tx_add_param(#name_literal, #name, &mut state)?;));
         }
     }
+    
+    let ident_literal = make_ident_literal_str(ident);
+    quote!(
+        let mut tr = self.tr.borrow_mut();
+        let mut state = tr.tx_begin_call(essrpc::MethodId{name: #ident_literal, num: #id})?;
+        #add_param_tokens
+        let state = tr.tx_finalize(&mut state)?;
+    )
+}
+
+
+fn impl_client_method(method: &TraitItemMethod, id: u32) -> TokenStream2 {
+    let ident = &method.sig.ident;
+    let param_tokens = &method.sig.decl.inputs;
+
+    if !verify_self_param_or_unneeded(method) {
+        return TokenStream2::new()
+    }
+
+    
     let rettype = match method.sig.decl.output {
         syn::ReturnType::Default => panic!("RPC methods must have a return type, {} does not ", ident),
         syn::ReturnType::Type(_arrow, ref t) => t
     };
-    let ident_literal = make_ident_literal_str(ident);
+
+    let tx_send = client_method_tx_send(method, id);
 
     quote!(
         fn #ident(#param_tokens) -> #rettype {
-            let mut tr = self.tr.borrow_mut();
-            let mut state = tr.tx_begin_call(essrpc::MethodId{name: #ident_literal, num: #id})?;
-            #add_param_tokens
-            tr.tx_finalize(&mut state)?;
+            #tx_send
             let ret: std::result::Result<#rettype, essrpc::RPCError> =
-                tr.rx_response(&mut state);
+                tr.rx_response(state);
             match ret {
                 Ok(v) => v,
                 Err(e) => Err(e.into())
@@ -107,24 +171,119 @@ fn impl_client_method(method: &TraitItemMethod, id: u32) -> TokenStream2 {
         })
 }
 
-fn create_client(trait_ident: &Ident, methods: &[TraitItemMethod]) -> TokenStream2 {
+
+fn impl_async_client_method(method: &TraitItemMethod, id: u32) -> TokenStream2 {
+    let ident = &method.sig.ident;
+    let param_tokens = &method.sig.decl.inputs;
+
+    if !verify_self_param_or_unneeded(method) {
+        return TokenStream2::new();
+    }
+
+    let rettype = get_future_return_type(method);
+    let errtype = get_error_type(method);
+    let tx_send = client_method_tx_send(method, id);
+
+    quote!(
+        fn #ident(#param_tokens) -> #rettype {
+            use futures::future::Future;
+            let send = || -> std::result::Result<TR::FinalState, #errtype> {
+                #tx_send
+                Ok(state)
+            };
+            match send() {
+                Err(e) => Box::new(futures::future::result(Err(e))),
+                Ok(state) => {
+                    let mut tr = self.tr.borrow_mut();
+                    Box::new(tr.rx_response(state).map_err(|e: essrpc::RPCError| e.into()))
+                }
+            }
+        })
+}
+
+fn create_async_client_trait(trait_ident: &Ident, methods: &[TraitItemMethod]) -> TokenStream2 {
+    let ident = async_client_trait_ident(trait_ident);
+    let mut method_decl_tokens = TokenStream2::new();
+    
+    for method in methods {
+        let mut new_method = method.clone();
+        let rettype = get_future_return_type(method);
+        new_method.sig.decl.output = parse_quote!(-> #rettype);
+        new_method.to_tokens(&mut method_decl_tokens);
+    }
+
+    quote!(
+        pub trait #ident {
+           #method_decl_tokens
+        }
+    )
+}
+
+fn get_future_return_type(method: &TraitItemMethod) -> syn::Type {
+    match get_result_types(&method.sig.decl.output) {
+            Some((ok_type, err_type)) => {
+                parse_quote!(Box<futures::Future<Item=#ok_type, Error=#err_type>>)
+            }
+            None => panic!("return {} type is not of expected form Result<T, E>",
+                           method.sig.decl.output.clone().into_token_stream())
+        }
+}
+
+fn get_error_type(method: &TraitItemMethod) -> syn::Type {
+    match get_result_types(&method.sig.decl.output) {
+            Some((_, err_type)) => err_type,
+            None => panic!("return {} type is not of expected form Result<T, E>",
+                           method.sig.decl.output.clone().into_token_stream())
+        }
+}
+
+/// For a return type of the form Result<T, E>, figure out what T and E are.
+fn get_result_types(result_type: &syn::ReturnType) -> Option<(syn::Type, syn::Type)> {
+    // unfortunately, there's no "if not let", meaning we end up with
+    // nesting instead of early return
+    if let syn::ReturnType::Type(_, b) = result_type {
+        if let syn::Type::Path(tp) = b.deref() {
+            let result_seg: syn::PathSegment = (*tp.path.segments.last()?.value()).clone();
+            if let syn::PathArguments::AngleBracketed(args) = result_seg.arguments {
+                let ok_type_generic = args.args.first()?.value().clone();
+                let err_type_generic = args.args.last()?.value().clone();
+                if let syn::GenericArgument::Type(ok_type) = ok_type_generic {
+                    if let syn::GenericArgument::Type(err_type) = err_type_generic {
+                        return Some((ok_type.clone(), err_type.clone()));
+                    }
+                }
+            }
+        }
+    }
+    None
+}
+
+fn create_client(trait_ident: &Ident, methods: &[TraitItemMethod], async_client: bool) -> TokenStream2 {
     let client_ident = client_ident(trait_ident);
+    let transport_ident = client_transport_ident(async_client);
+    let rpcclient_ident = rpcclient_ident(async_client);
 
     let mut method_impl_tokens = TokenStream2::new();
 
     let mut mcnt = 0;
     for method in methods {
-        method_impl_tokens.extend(impl_client_method(method, mcnt));
+        method_impl_tokens.extend(
+            if async_client {
+                impl_async_client_method(method, mcnt)
+            } else {
+                impl_client_method(method, mcnt)
+            }
+        );
         mcnt += 1;
     }
 
     quote!(
-        pub struct #client_ident<TR: essrpc::ClientTransport> {
+        pub struct #client_ident<TR: essrpc::#transport_ident> {
             tr: std::cell::RefCell<TR>,
         }
 
-        impl <TR> essrpc::RPCClient for #client_ident<TR> where
-            TR: essrpc::ClientTransport {
+        impl <TR> essrpc::#rpcclient_ident for #client_ident<TR> where
+            TR: essrpc::#transport_ident {
 
             type TR = TR;
             
@@ -134,7 +293,7 @@ fn create_client(trait_ident: &Ident, methods: &[TraitItemMethod]) -> TokenStrea
         }
 
         impl <TR> #trait_ident for #client_ident<TR> where
-            TR: essrpc::ClientTransport {
+            TR: essrpc::#transport_ident {
             
             #method_impl_tokens
         }